feat(ui): add scroll to palette tray
Added scroll function to the palette tray in the toolbar. The user can
hold the left mouse button and drag the palette tray left or right to
view the rest of the colors.

Tiny fix: user can no longer select the color index text of a color
sample in the palette tray.

Resolves: #4